+++ CHANGES.md
@@ -1,5 +1,12 @@
 # Changes
 
+## 0.0.4 (2020-07-10)
+
+- Added Python 3.8 support
+- Dropped Python 3.4 support
+- FIX: Critical bug in error handling, see #26 and #27.
+- Improved library structure.
+
 ## 0.0.3 (2019-05-18)
 
 - Removed Python-2-compatibility code